fixup! User-defined literal suffixes: Tests

+++ test/libsolidity/syntaxTests/literalSuffixes/parameters/overloading_decimal_vs_decimal.sol
@@ -1,19 +1,19 @@
-function uSuffix(uint8, uint) pure suffix returns (int8) {}
-function uSuffix(uint16, uint) pure suffix returns (int16) {}
+function uSuffix(uint8, uint) pure suffix returns (uint) {}
+function uSuffix(uint16, uint) pure suffix returns (int) {}
 
-function iSuffix(int8, uint) pure suffix returns (int24) {}
-function iSuffix(int16, uint) pure suffix returns (int32) {}
+function iSuffix(int8, uint) pure suffix returns (uint) {}
+function iSuffix(int16, uint) pure suffix returns (int) {}
 
-function iuSuffix(uint8, uint) pure suffix returns (int40) {}
-function iuSuffix(int8, uint) pure suffix returns (int48) {}
+function iuSuffix(uint8, uint) pure suffix returns (int) {}
+function iuSuffix(int8, uint) pure suffix returns (uint) {}
 
 contract C {
-    int16 a = 1.024 uSuffix;
-    int32 b = 1.024 iSuffix;
+    int a = 1.024 uSuffix;
+    int b = 1.024 iSuffix;
 
-    int16 c = -1.024 uSuffix;
-    int32 d = -1.024 iSuffix;
+    int c = -1.024 uSuffix;
+    int d = -1.024 iSuffix;
 
-    int40 e = 2.55 iuSuffix;
-    int40 f = -2.55 iuSuffix;
+    int e = 2.55 iuSuffix;
+    int f = -2.55 iuSuffix;
 }
