+++ test/libsolidity/syntaxTests/literalSuffixes/parameters/overloading_integer_vs_integer.sol
@@ -1,19 +1,19 @@
-function uSuffix(uint8) pure suffix returns (int8) {}
-function uSuffix(uint16) pure suffix returns (int16) {}
+function uSuffix(uint8) pure suffix returns (uint) {}
+function uSuffix(uint16) pure suffix returns (int) {}
 
-function iSuffix(int8) pure suffix returns (int24) {}
-function iSuffix(int16) pure suffix returns (int32) {}
+function iSuffix(int8) pure suffix returns (uint) {}
+function iSuffix(int16) pure suffix returns (int) {}
 
-function iuSuffix(uint8) pure suffix returns (int40) {}
-function iuSuffix(int8) pure suffix returns (int48) {}
+function iuSuffix(uint8) pure suffix returns (int) {}
+function iuSuffix(int8) pure suffix returns (uint) {}
 
 contract C {
-    int16 a = 1024 uSuffix;
-    int32 b = 1024 iSuffix;
+    int a = 1024 uSuffix;
+    int b = 1024 iSuffix;
 
-    int16 c = -1024 uSuffix;
-    int32 d = -1024 iSuffix;
+    int c = -1024 uSuffix;
+    int d = -1024 iSuffix;
 
-    int40 e = 255 iuSuffix;
-    int40 f = -255 iuSuffix;
+    int e = 255 iuSuffix;
+    int f = -255 iuSuffix;
 }
